fix(test): satisfy repository contract guardrails
